+++ Phase-5/SearchEngine/Classes/InvertedIndex.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using SearchEngine.Interfaces;
 
-namespace SearchEngine
+namespace SearchEngine.Classes
 {
     public class InvertedIndex : IInvertedIndex
     {
