+++ Phase-5/SearchEngine/Classes/SearchEngineCore.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using SearchEngine.Interfaces;
 
-namespace SearchEngine
+namespace SearchEngine.Classes
 {
     public class SearchEngineCore : ISearchEngineCore
     {
