+++ Phase-5/SearchEngine/Interfaces/IManager.cs
@@ -1,14 +1,18 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace SearchEngine.Interfaces
 {
     public interface IManager
     {
+        static void PrintElements(ICollection<int> elements)
+        {
+            foreach (int id in elements)
+            {
+                Console.WriteLine("element" + id);
+            }
+        }
         public void Run();
-        void MakeSearchEngine();
-        void MakeInvertedIndex(string path);
-        HashSet<int> DoSearch(string toSearch);
-        void PrintElements(ICollection<int> elements);
         bool Finished(string toSearch);
     }
 }
