+++ Phase-5/SearchEngineTests/InvertedIndexTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using NSubstitute;
 using SearchEngine;
+using SearchEngine.Classes;
 using SearchEngine.Interfaces;
 using Xunit;
