+++ Phase-5/SearchEngineTests/SearchEngineCoreTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using NSubstitute;
 using SearchEngine;
+using SearchEngine.Classes;
 using SearchEngine.Interfaces;
 using Xunit;
