+++ Phase-5/SearchEngineTests/WordProcessorTests.cs
@@ -1,4 +1,5 @@
 ﻿using SearchEngine;
+using SearchEngine.Classes;
 using Xunit;
 
 namespace SearchEngineTests
